# api/openapi.yaml
openapi: 3.1.0
info:
  title: タスク管理システム API
  description: |
    マルチテナントSaaS型タスク管理システムのREST API仕様(基本設計書 v1.5.1 / 要件定義書 v1.4.4 準拠)。

    認証はKeycloak経由のOIDC、本APIはJWTを Spring Security OAuth2 Resource Server で検証する。
    全リクエストには `Authorization: Bearer {access_token}` を付与し、
    業務 API ではアクティブテナントを `X-Tenant-Id` ヘッダで送出する。
    プラットフォーム運用 API(`/api/auth/*` の一部・`/api/tenants` 系 SaaS Admin 操作・セルフサインアップ・`/api/platform/*`)は `X-Tenant-Id` を要求しない。

    認可は**プラットフォーム軸**と**テナント軸**の 2 軸で直交する(参照のみの Viewer は設けない)。
    タスクごとの参照認可は visibility (TENANT/STAKEHOLDERS/PRIVATE) と所有者・担当者・関係者の関係で決定する(ADR-0005 / 基本設計書 §6.2.1)。

    ## 認可モデル詳細(2 軸直交)

    - **プラットフォーム軸 / SaaS Admin**: Keycloak realm role `APP_ADMIN` で管理する(本システム DB には保持しない)。
      テナント運営 API(`/api/tenants/**` 状態管理・`/api/platform/**` 監視)のみ操作可能。**業務 API(`/api/tasks/**`、`/api/dashboard/**`、`/api/tenant/users/**`、`/api/tenant/dashboard/**`(Tenant Admin 限定)、`/api/audit-logs`(Tenant Admin 限定)等)には触れない**(呼び出された場合は 403)。Spring Security 上は `hasRole('APP_ADMIN')`。
    - **テナント軸 / Tenant Admin**: 本システム DB の `user_tenants.role = TENANT_ADMIN` で管理。
      **テナント運営権限のみ**(ユーザー招待、ロール管理、監査ログ参照、テナント運営者向けダッシュボード `/api/tenant/dashboard/summary` 参照)。**業務タスクへの特別権限は持たない**(ADR-0005)。タスク操作は所有者・担当者・関係者の役割として通常ルール(下記)で行う。Spring Security 上は `hasRole('TENANT_ADMIN')`。
    - **テナント軸 / Member**: 本システム DB の `user_tenants.role = MEMBER` で管理。
...

...

    ## タスク認可ルール(ADR-0005 / 基本設計書 §6.2.1)

    タスク認可は **所有者(Owner)・担当者(Assignee)・関係者(Stakeholder)の 3 役割のみで評価**する。

...
    - 公開範囲変更: 所有者のみ
...
servers:
...
security:
  - bearerAuth: []
    tenantHeader: []

# ======================================================================
# パス
# ======================================================================
paths:
...
  /api/invitations/{token}:
    get:
      tags: [User]
      summary: 招待照会 (受諾画面用・公開)
      description: |
        認可: なし(公開)。受諾画面表示用にトークンから招待詳細(テナント名・email・ロール・受諾可否状態)を返す。**トークンは消費しない**。
        トークンが参加先テナントを保持するため `X-Tenant-Id` 不要(ADR-0040 §3.3)。
        期限切れ / 使用済み / 失効は `status` に状態を載せて 200 を返し(404 にしない)、画面側で案内分岐する。
        トークンに対応する招待が無い場合のみ 404(存在有無を漏らさない)。
      operationId: getInvitation
      security: []
      parameters:
        - { name: token, in: path, required: true, description: "受諾トークン(平文)", schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InvitationDetail" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [User]
      summary: 招待受諾 (公開)
      description: |
        認可: なし(公開)。トークンを消費して招待を受諾する(ADR-0040 §3.3)。
        未登録 email は会員登録(profile + Keycloak credential)+ `user_tenants` 紐付け + 招待 USED を確定する。
        既に登録済みの email は会員登録不要のため 409 を返し「ログインして参加」へ誘導する(認証済みの参加紐付けは後続フロー)。
        期限切れ / 使用済み / 失効は 409。トークン不在は 404。
      operationId: acceptInvitation
      security: []
      parameters:
        - { name: token, in: path, required: true, description: "受諾トークン(平文)", schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AcceptInvitationRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AcceptInvitationResponse" }
        "400": { $ref: "#/components/responses/Validation" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          $ref: "#/components/responses/Conflict"
          description: 期限切れ / 使用済み / 失効、または既に登録済み(ログインして参加)。
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: KeycloakがOIDCで発行するJWT
    tenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-Id
      description: アクティブテナントのID
  responses:
...
    NotFound:
      description: リソース未存在、または参照権限不足(リソース存在を漏らさない方針)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Validation:
      description: 入力検証エラー
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: リソースの競合状態
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [timestamp, status, code, message, path]
      properties:
        timestamp: { type: string, format: date-time }
        status:    { type: integer, example: 400 }
        error:     { type: string, example: "Bad Request" }
        code:      { type: string, example: "E_VALIDATION", description: "業務エラーコード" }
        message:   { type: string }
        path:      { type: string }
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:   { type: string }
              message: { type: string }

    # --- 列挙 ---
...
    Role:
      type: string
      enum: [TENANT_ADMIN, MEMBER]
...
    InvitationDisplayStatus:
      type: string
      description: 受諾画面の状態。`PENDING` のみ受諾可能。
      enum: [PENDING, EXPIRED, USED, REVOKED]
    InvitationDetail:
      type: object
      required: [email, tenantName, role, status, alreadyRegistered]
      properties:
        email:            { type: string, format: email, description: 招待先メールアドレス }
        tenantName:       { type: string, description: 参加先テナント名 }
        role:             { $ref: "#/components/schemas/Role" }
        status:           { $ref: "#/components/schemas/InvitationDisplayStatus" }
        alreadyRegistered: { type: boolean, description: "true ならログインして参加へ誘導" }
    AcceptInvitationRequest:
      type: object
      required: [fullName, fullNameKana, password]
      description: 未登録ユーザーの会員登録に使う profile + password。email は招待トークンが保持するため受け取らない。
      properties:
        fullName:       { type: string, maxLength: 255 }
        fullNameKana:   { type: string, maxLength: 255 }
        departmentName: { type: string, maxLength: 255, nullable: true }
        password:       { type: string, minLength: 8, maxLength: 255 }
    AcceptInvitationResponse:
      type: object
      required: [userId, tenantId]
      properties:
        userId:   { type: integer, format: int64, description: 参加したユーザーID }
        tenantId: { type: integer, format: int64, description: 参加先テナントID }

    # --- タスク ---
...